Add files via upload
A Kenyan entrepreneur has created an online cryptography course and would want to advertise it on her blog. She currently targets audiences originating from various countries. In the past, she ran ads to advertise a related course on the same blog and collected data in the process. She would now like to employ the services of a Data Science Consultant to help her identify which individuals are most likely to click on her ads.